src/requirements.txt
aniso8601==6.0.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.4
django-graphiql==0.4.4
django-graphql-jwt==0.2.2
entrypoints==0.3
flake8==3.7.8
graphene==2.1.7
graphene-django==2.5.0
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
importlib-metadata==0.20
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
promise==2.2.1
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pytest==5.1.2
pytest-django==3.5.1
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
